Rename: pages폴더 login.tsx & signup.tsx & slug.tsx 각 폴더의 index.tsx로 변경
- 라우트 방식을 파일이름이 아닌 폴더방식으로 변경
