+++ pages/[slug]/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const index = () => {
+  return <div>index</div>;
+};
+
+export default index;
