+++ pages/mypage/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Mypage = () => {
+  return <div>Mypage</div>;
+};
+
+export default Mypage;
